+++ LadybugTests/AppleRadarTests.swift
@@ -23,14 +23,14 @@ class AppleRadarTests: XCTestCase {
 
     func testParseURL() {
         let url = URL(string: "https://bugreport.apple.com/web/?problemID=41075977")!
-        let radarID = AppleRadar.parse(url)
-        XCTAssertNotNil(radarID)
-        XCTAssert((radarID?.id == 41075977) == true)
+        let radarNumber = AppleRadarURL.parse(url)
+        XCTAssertNotNil(radarNumber)
+        XCTAssert((radarNumber?.rawValue == 41075977) == true)
     }
     
     func testBuildURL() {
-        let radarID = RadarID(12345678)
-        let url = AppleRadar.buildURL(from: radarID)
+        let radarID = RadarNumber(12345678)
+        let url = AppleRadarURL.buildURL(from: radarID)
         XCTAssertEqual(url.absoluteString, "https://bugreport.apple.com/web/?problemID=12345678")
     }
 }